requirements.txt
fastmcp>=0.1.0
pillow>=10.0.0
pygetwindow>=0.0.9
requests>=2.31.0
